test: fuzzer mode for AFL's persistent mode
When fuzzing with AFL, if the LLVM-based instrumentation is
used (via the afl-clang-fast wrapper), then it is possible to
have a single execution of the fuzzer program iterate multiple
times over the fuzzing entrypoint (similar to libFuzzer's normal
mode of execution) with different data.  This is much (e.g. 10x)
faster.

Add code to support this, by checking whether __AFL_LOOP is
defined at compile-time.

Also, shift the code to effectively be C rather than C++.