Added support for Windows DNS Suffix Search List (#93)
This change solves issue #53.

Support for suffix search lists was already built in for Linux. The search list could be set via set_search. With this change the suffix search list from Windows is read from the registry and then set into the ares configuration via set_search. There are two sources for the search list:

The global DNS suffix search list.
The primary and connection specific DNS suffixes if the global is not available.

Contributed by @ChristianAmmer