Android variants may not have __system_property_get
Some android systems like ARM64 may not have the __system_property_get
symbol in libc (but still have it in the public headers).  Detect this
condition at build time.  The __system_property_get method of retrieving
name servers is deprecated as of Oreo so should strictly be a fallback
mechanism anyhow.