gethostbyname: reload resolv.conf values if file changed
There is unhandled ares use case when we call ares_init when /etc/resolv.conf
is empty and later do the DNS request when it is filled in.
The bug is that ares does not detect change of resolv.conf.
The result is that default DNS IP is used, which is 127.0.0.1

Solution for this problem is that we check if resolv.conf file changed
each time we call ares_gethostbyname function.
Similar solution is done in libc.

Another unhandled use case is that system changes resolve.conf for some reason
and ares would not be aware of this change silently doing wrong DNS requests.

Signed-off-by: Nayana Topolsky <nayana.topolsky@streamunlimited.com>