CMake: Rework library function checking (#310)
CHECK_LIBRARY_EXISTS(), while it takes a function name, does not actually verify the function exists in the library being evaluated. Instead, if the function is found in any dependent library, and the referenced library also exists, it returns true. This is not desirable.

Wrap with a Macro to change the behavior.

Fixes: #307
Fix By: Brad House (@bradh352)