ares__sortaddrinfo: fix crash if the host list is empty
The bug was generated because there was not check for the number
of items in the list and invalid memory was accesed when the list
was empty. There is a check for null after calling malloc but on
some systems it always returns a valid address for size equals 0.

For e.g., on macOS and Windows systems malloc returns a valid address:

https://man.openbsd.org/malloc.3
https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/malloc

However, on Linux malloc returns NULL if size equals 0:

https://man7.org/linux/man-pages/man3/malloc.3.html

Also, the openBSD manual says it more clearly:

If nmemb or size is equal to 0, a unique pointer to an access protected,
zero sized object is returned. Access via this pointer will generate a
SIGSEGV exception.

Relates To: #392, 0903dcecabca283d0fa771632892dc7592b7a66d