Watcom Portability Improvements
- Modify the Watcom Makefile for the source code reorganization (#352)
- Add *.map files into .gitignore
- Fix build errors with Watcom's builtin Windows SDK (which is rather
  outdated). It's smart enough to understand Windows Vista, but doesn't
  have PMIB_UNICASTIPADDRESS_TABLE or MIB_IPFORWARD_ROW2.

It may be possible to use a different Windows SDK with the Watcom
compiler, such as the most recent Windows 10 SDK. Alternatively the SDK
in OpenWatcom 2.0 (which is in development) should fix this.

I have no problems testing this Makefile prior to releases, just give me
a ping.

Tested with Windows Vista, Windows 7, and Windows 10 using 'adig',
'acountry', and 'ahost'. This also seems to work on Windows XP, though
this is likely due to the compiler in use.