Support configuration of DNS server ports
As per https://man.openbsd.org/OpenBSD-5.1/resolv.conf.5 we should
suppor bracketed syntax for resolv.conf entries to contain an optional
port number.

We also need to utilize this format for configuration of MacOS
DNS servers as seen when using the Viscosity OpenVPN client, where
it starts a private DNS server listening on localhost on a non-standard
port.