Fix building c-ares-1.22.0 and higher under Watcom.
This commit fixes building c-ares using the WATCOM compiler by adding
-dHAS_STDINT_H into the CFLAGS. On a configure or CMake based
configuration, this would be added in by that process.

This fixes a build error when compiling ares__buf.c, due to not having
SIZE_MAX declared.

Tested by linking to the library as well as using adig and ahost.

Fixes #622